エネミーマネージャーの追加とエネミースポナーの削除
新しいエネミーマネージャークラスをEnemyManager.cppとEnemyManager.hに実装し、敵の生成、更新、描画、コライダーの登録、ImGui描画処理、ウェーブの開始機能を追加しました。
EnemySpawner.cppとEnemySpawner.hからはエネミースポナーの機能を削除し、エネミーマネージャーに統合しました。
GamePlayScene.cppとGamePlayScene.hでは、エネミースポナーのインスタンスをエネミーマネージャーに置き換え、関連するメンバ変数も更新しました。
