+++ Project/ApplicationLayer/Enemy/EnemyManager.h
@@ -1,0 +1,54 @@
+#pragma once
+#include "Enemy.h"
+#include <memory>
+#include <vector>
+
+
+/// ---------- 前方宣言 ---------- ///
+class Player;
+
+
+/// -------------------------------------------------------------
+///				　		エネミーマネージャー
+/// -------------------------------------------------------------
+class EnemyManager
+{
+public: /// ---------- メンバ関数 ---------- ///
+
+	// 初期化処理
+	void Initialize(Player* player);
+
+	// 更新処理
+	void Update();
+
+	// 描画処理
+	void Draw();
+
+	void RegisterColliders(class CollisionManager* collisionManager);
+
+	// ImGui描画処理
+	void DrawImGui();
+
+	// 次のウェーブを開始
+	void StartNextWave();
+
+	// ウェーブクリア判定
+	bool IsWaveClear() const { return spawnedEnemies_ >= totalEnemiesThisWave_ && enemies_.empty(); }
+
+	// 現在のウェーブ数を取得
+	int GetCurrentWave() const { return currentWave_; }
+
+	// スポーンしたエネミーの数を取得
+	const std::vector<std::unique_ptr<Enemy>>& GetEnemies() const { return enemies_; }
+
+private: /// ---------- メンバ変数 ---------- ///
+
+	std::vector<std::unique_ptr<Enemy>> enemies_; // エネミーのリスト
+	Player* player_ = nullptr; // プレイヤーへのポインタ
+
+	int currentWave_ = 1;		   // 現在のウェーブ数
+	int totalEnemiesThisWave_ = 0; // 現在のウェーブでスポーンするエネミーの数
+	int spawnedEnemies_ = 0;	   // スポーンしたエネミーの数
+	float spawnInterval_ = 1.0f;   // スポーン間隔
+	float spawnTimer_ = 0.0f;	   // スポーンタイマー
+};
